@@ -17,7 +17,7 @@ version: 0.0.0+000000000
 app_version: 0.0.0
 
 environment:
-  sdk: '>=3.3.0 <4.0.0'
+  sdk: ^3.8.0
   flutter: ^3.19.0
 
 dependencies:
@@ -53,7 +53,7 @@ dependencies:
   freezed_annotation: ^3.1.0
   google_fonts: ^6.2.1
   intl: any
-  json_annotation: ^4.8.1
+  json_annotation: ^4.9.0
   just_audio: ^0.10.1
   linkify: ^5.0.0
   logging: ^1.2.0
@@ -128,7 +128,7 @@ dev_dependencies:
   flutter_lints: ^4.0.0
   flutter_native_splash: ^2.4.5
   freezed: ^3.2.3
-  json_serializable: ^6.7.1
+  json_serializable: ^6.11.1
   patrol: ^3.15.1
   mocktail: ^1.0.4
 
