@@ -1,6 +1,11 @@
 part of 'notifications_bloc.dart';
 
 @freezed
-abstract class NotificationsState with _$NotificationsState {
-  const factory NotificationsState([Notification? lastNotification]) = _NotificationsState;
+class NotificationsState with _$NotificationsState {
+  const NotificationsState({
+    this.lastNotification,
+  });
+
+  @override
+  final Notification? lastNotification;
 }
