@@ -11,1543 +11,1177 @@ part of 'app_bloc.dart';
 
 // dart format off
 T _$identity<T>(T value) => value;
-
 /// @nodoc
 mixin _$SessionUpdated {
-  Session? get session;
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _SessionUpdated &&
-            (identical(other.session, session) || other.session == session));
-  }
+ Session? get session;
 
-  @override
-  int get hashCode => Object.hash(runtimeType, session);
 
-  @override
-  String toString() {
-    return '_SessionUpdated(session: $session)';
-  }
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _SessionUpdated&&(identical(other.session, session) || other.session == session));
 }
+
+
+@override
+int get hashCode => Object.hash(runtimeType,session);
+
+@override
+String toString() {
+  return '_SessionUpdated(session: $session)';
+}
+
+
+}
+
+
+
 
 /// Adds pattern-matching-related methods to [_SessionUpdated].
 extension _SessionUpdatedPatterns on _SessionUpdated {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(__SessionUpdated value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case __SessionUpdated() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>(TResult Function( __SessionUpdated value)?  $default,{required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case __SessionUpdated() when $default != null:
+return $default(_that);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(__SessionUpdated value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case __SessionUpdated():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult map<TResult extends Object?>(TResult Function( __SessionUpdated value)  $default,){
+final _that = this;
+switch (_that) {
+case __SessionUpdated():
+return $default(_that);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(__SessionUpdated value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case __SessionUpdated() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(TResult? Function( __SessionUpdated value)?  $default,){
+final _that = this;
+switch (_that) {
+case __SessionUpdated() when $default != null:
+return $default(_that);case _:
+  return null;
 
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(Session? session)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case __SessionUpdated() when $default != null:
-        return $default(_that.session);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( Session? session)?  $default,{required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case __SessionUpdated() when $default != null:
+return $default(_that.session);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(Session? session) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case __SessionUpdated():
-        return $default(_that.session);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( Session? session)  $default,) {final _that = this;
+switch (_that) {
+case __SessionUpdated():
+return $default(_that.session);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(Session? session)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case __SessionUpdated() when $default != null:
-        return $default(_that.session);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( Session? session)?  $default,) {final _that = this;
+switch (_that) {
+case __SessionUpdated() when $default != null:
+return $default(_that.session);case _:
+  return null;
+
+}
+}
+
 }
 
 /// @nodoc
+
 
 class __SessionUpdated implements _SessionUpdated {
   const __SessionUpdated(this.session);
+  
 
-  @override
-  final Session? session;
+@override final  Session? session;
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is __SessionUpdated &&
-            (identical(other.session, session) || other.session == session));
-  }
 
-  @override
-  int get hashCode => Object.hash(runtimeType, session);
 
-  @override
-  String toString() {
-    return '_SessionUpdated(session: $session)';
-  }
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is __SessionUpdated&&(identical(other.session, session) || other.session == session));
 }
+
+
+@override
+int get hashCode => Object.hash(runtimeType,session);
+
+@override
+String toString() {
+  return '_SessionUpdated(session: $session)';
+}
+
+
+}
+
+
+
 
 /// @nodoc
 mixin _$AppThemeSettingsChanged {
-  ThemeSettings get value;
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is AppThemeSettingsChanged &&
-            (identical(other.value, value) || other.value == value));
-  }
+ ThemeSettings get value;
 
-  @override
-  int get hashCode => Object.hash(runtimeType, value);
 
-  @override
-  String toString() {
-    return 'AppThemeSettingsChanged(value: $value)';
-  }
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is AppThemeSettingsChanged&&(identical(other.value, value) || other.value == value));
 }
+
+
+@override
+int get hashCode => Object.hash(runtimeType,value);
+
+@override
+String toString() {
+  return 'AppThemeSettingsChanged(value: $value)';
+}
+
+
+}
+
+
+
 
 /// Adds pattern-matching-related methods to [AppThemeSettingsChanged].
 extension AppThemeSettingsChangedPatterns on AppThemeSettingsChanged {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_AppThemeSettingsChanged value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _AppThemeSettingsChanged() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>(TResult Function( _AppThemeSettingsChanged value)?  $default,{required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _AppThemeSettingsChanged() when $default != null:
+return $default(_that);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_AppThemeSettingsChanged value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _AppThemeSettingsChanged():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult map<TResult extends Object?>(TResult Function( _AppThemeSettingsChanged value)  $default,){
+final _that = this;
+switch (_that) {
+case _AppThemeSettingsChanged():
+return $default(_that);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_AppThemeSettingsChanged value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _AppThemeSettingsChanged() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(TResult? Function( _AppThemeSettingsChanged value)?  $default,){
+final _that = this;
+switch (_that) {
+case _AppThemeSettingsChanged() when $default != null:
+return $default(_that);case _:
+  return null;
 
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(ThemeSettings value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _AppThemeSettingsChanged() when $default != null:
-        return $default(_that.value);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( ThemeSettings value)?  $default,{required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _AppThemeSettingsChanged() when $default != null:
+return $default(_that.value);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(ThemeSettings value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _AppThemeSettingsChanged():
-        return $default(_that.value);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( ThemeSettings value)  $default,) {final _that = this;
+switch (_that) {
+case _AppThemeSettingsChanged():
+return $default(_that.value);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(ThemeSettings value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _AppThemeSettingsChanged() when $default != null:
-        return $default(_that.value);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( ThemeSettings value)?  $default,) {final _that = this;
+switch (_that) {
+case _AppThemeSettingsChanged() when $default != null:
+return $default(_that.value);case _:
+  return null;
+
+}
+}
+
 }
 
 /// @nodoc
+
 
 class _AppThemeSettingsChanged implements AppThemeSettingsChanged {
   const _AppThemeSettingsChanged(this.value);
+  
 
-  @override
-  final ThemeSettings value;
+@override final  ThemeSettings value;
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _AppThemeSettingsChanged &&
-            (identical(other.value, value) || other.value == value));
-  }
 
-  @override
-  int get hashCode => Object.hash(runtimeType, value);
 
-  @override
-  String toString() {
-    return 'AppThemeSettingsChanged(value: $value)';
-  }
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _AppThemeSettingsChanged&&(identical(other.value, value) || other.value == value));
 }
+
+
+@override
+int get hashCode => Object.hash(runtimeType,value);
+
+@override
+String toString() {
+  return 'AppThemeSettingsChanged(value: $value)';
+}
+
+
+}
+
+
+
 
 /// @nodoc
 mixin _$AppThemeModeChanged {
-  ThemeMode get value;
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is AppThemeModeChanged &&
-            (identical(other.value, value) || other.value == value));
-  }
+ ThemeMode get value;
 
-  @override
-  int get hashCode => Object.hash(runtimeType, value);
 
-  @override
-  String toString() {
-    return 'AppThemeModeChanged(value: $value)';
-  }
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is AppThemeModeChanged&&(identical(other.value, value) || other.value == value));
 }
+
+
+@override
+int get hashCode => Object.hash(runtimeType,value);
+
+@override
+String toString() {
+  return 'AppThemeModeChanged(value: $value)';
+}
+
+
+}
+
+
+
 
 /// Adds pattern-matching-related methods to [AppThemeModeChanged].
 extension AppThemeModeChangedPatterns on AppThemeModeChanged {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_AppThemeModeChanged value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _AppThemeModeChanged() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>(TResult Function( _AppThemeModeChanged value)?  $default,{required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _AppThemeModeChanged() when $default != null:
+return $default(_that);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_AppThemeModeChanged value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _AppThemeModeChanged():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult map<TResult extends Object?>(TResult Function( _AppThemeModeChanged value)  $default,){
+final _that = this;
+switch (_that) {
+case _AppThemeModeChanged():
+return $default(_that);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_AppThemeModeChanged value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _AppThemeModeChanged() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(TResult? Function( _AppThemeModeChanged value)?  $default,){
+final _that = this;
+switch (_that) {
+case _AppThemeModeChanged() when $default != null:
+return $default(_that);case _:
+  return null;
 
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(ThemeMode value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _AppThemeModeChanged() when $default != null:
-        return $default(_that.value);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( ThemeMode value)?  $default,{required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _AppThemeModeChanged() when $default != null:
+return $default(_that.value);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(ThemeMode value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _AppThemeModeChanged():
-        return $default(_that.value);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( ThemeMode value)  $default,) {final _that = this;
+switch (_that) {
+case _AppThemeModeChanged():
+return $default(_that.value);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(ThemeMode value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _AppThemeModeChanged() when $default != null:
-        return $default(_that.value);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( ThemeMode value)?  $default,) {final _that = this;
+switch (_that) {
+case _AppThemeModeChanged() when $default != null:
+return $default(_that.value);case _:
+  return null;
+
+}
+}
+
 }
 
 /// @nodoc
+
 
 class _AppThemeModeChanged implements AppThemeModeChanged {
   const _AppThemeModeChanged(this.value);
+  
 
-  @override
-  final ThemeMode value;
+@override final  ThemeMode value;
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _AppThemeModeChanged &&
-            (identical(other.value, value) || other.value == value));
-  }
 
-  @override
-  int get hashCode => Object.hash(runtimeType, value);
 
-  @override
-  String toString() {
-    return 'AppThemeModeChanged(value: $value)';
-  }
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _AppThemeModeChanged&&(identical(other.value, value) || other.value == value));
 }
+
+
+@override
+int get hashCode => Object.hash(runtimeType,value);
+
+@override
+String toString() {
+  return 'AppThemeModeChanged(value: $value)';
+}
+
+
+}
+
+
+
 
 /// @nodoc
 mixin _$AppLocaleChanged {
-  Locale get value;
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is AppLocaleChanged &&
-            (identical(other.value, value) || other.value == value));
-  }
+ Locale get value;
 
-  @override
-  int get hashCode => Object.hash(runtimeType, value);
 
-  @override
-  String toString() {
-    return 'AppLocaleChanged(value: $value)';
-  }
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is AppLocaleChanged&&(identical(other.value, value) || other.value == value));
 }
+
+
+@override
+int get hashCode => Object.hash(runtimeType,value);
+
+@override
+String toString() {
+  return 'AppLocaleChanged(value: $value)';
+}
+
+
+}
+
+
+
 
 /// Adds pattern-matching-related methods to [AppLocaleChanged].
 extension AppLocaleChangedPatterns on AppLocaleChanged {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_AppLocaleChanged value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _AppLocaleChanged() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>(TResult Function( _AppLocaleChanged value)?  $default,{required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _AppLocaleChanged() when $default != null:
+return $default(_that);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_AppLocaleChanged value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _AppLocaleChanged():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult map<TResult extends Object?>(TResult Function( _AppLocaleChanged value)  $default,){
+final _that = this;
+switch (_that) {
+case _AppLocaleChanged():
+return $default(_that);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_AppLocaleChanged value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _AppLocaleChanged() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(TResult? Function( _AppLocaleChanged value)?  $default,){
+final _that = this;
+switch (_that) {
+case _AppLocaleChanged() when $default != null:
+return $default(_that);case _:
+  return null;
 
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(Locale value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _AppLocaleChanged() when $default != null:
-        return $default(_that.value);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( Locale value)?  $default,{required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _AppLocaleChanged() when $default != null:
+return $default(_that.value);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(Locale value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _AppLocaleChanged():
-        return $default(_that.value);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( Locale value)  $default,) {final _that = this;
+switch (_that) {
+case _AppLocaleChanged():
+return $default(_that.value);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(Locale value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _AppLocaleChanged() when $default != null:
-        return $default(_that.value);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( Locale value)?  $default,) {final _that = this;
+switch (_that) {
+case _AppLocaleChanged() when $default != null:
+return $default(_that.value);case _:
+  return null;
+
+}
+}
+
 }
 
 /// @nodoc
+
 
 class _AppLocaleChanged implements AppLocaleChanged {
   const _AppLocaleChanged(this.value);
+  
 
-  @override
-  final Locale value;
+@override final  Locale value;
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _AppLocaleChanged &&
-            (identical(other.value, value) || other.value == value));
-  }
 
-  @override
-  int get hashCode => Object.hash(runtimeType, value);
 
-  @override
-  String toString() {
-    return 'AppLocaleChanged(value: $value)';
-  }
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _AppLocaleChanged&&(identical(other.value, value) || other.value == value));
 }
+
+
+@override
+int get hashCode => Object.hash(runtimeType,value);
+
+@override
+String toString() {
+  return 'AppLocaleChanged(value: $value)';
+}
+
+
+}
+
+
+
 
 /// @nodoc
 mixin _$AppAgreementAccepted {
-  AgreementStatus get status;
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is AppAgreementAccepted &&
-            (identical(other.status, status) || other.status == status));
-  }
+ AgreementStatus get status;
 
-  @override
-  int get hashCode => Object.hash(runtimeType, status);
 
-  @override
-  String toString() {
-    return 'AppAgreementAccepted(status: $status)';
-  }
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is AppAgreementAccepted&&(identical(other.status, status) || other.status == status));
 }
+
+
+@override
+int get hashCode => Object.hash(runtimeType,status);
+
+@override
+String toString() {
+  return 'AppAgreementAccepted(status: $status)';
+}
+
+
+}
+
+
+
 
 /// Adds pattern-matching-related methods to [AppAgreementAccepted].
 extension AppAgreementAcceptedPatterns on AppAgreementAccepted {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(_UserAppAgreementUpdate value)? updateUserAgreement,
-    TResult Function(_ContactsAppAgreementUpdate value)?
-        updateContactsAgreement,
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _UserAppAgreementUpdate() when updateUserAgreement != null:
-        return updateUserAgreement(_that);
-      case _ContactsAppAgreementUpdate() when updateContactsAgreement != null:
-        return updateContactsAgreement(_that);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>({TResult Function( _UserAppAgreementUpdate value)?  updateUserAgreement,TResult Function( _ContactsAppAgreementUpdate value)?  updateContactsAgreement,required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _UserAppAgreementUpdate() when updateUserAgreement != null:
+return updateUserAgreement(_that);case _ContactsAppAgreementUpdate() when updateContactsAgreement != null:
+return updateContactsAgreement(_that);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(_UserAppAgreementUpdate value)
-        updateUserAgreement,
-    required TResult Function(_ContactsAppAgreementUpdate value)
-        updateContactsAgreement,
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _UserAppAgreementUpdate():
-        return updateUserAgreement(_that);
-      case _ContactsAppAgreementUpdate():
-        return updateContactsAgreement(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult map<TResult extends Object?>({required TResult Function( _UserAppAgreementUpdate value)  updateUserAgreement,required TResult Function( _ContactsAppAgreementUpdate value)  updateContactsAgreement,}){
+final _that = this;
+switch (_that) {
+case _UserAppAgreementUpdate():
+return updateUserAgreement(_that);case _ContactsAppAgreementUpdate():
+return updateContactsAgreement(_that);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(_UserAppAgreementUpdate value)? updateUserAgreement,
-    TResult? Function(_ContactsAppAgreementUpdate value)?
-        updateContactsAgreement,
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _UserAppAgreementUpdate() when updateUserAgreement != null:
-        return updateUserAgreement(_that);
-      case _ContactsAppAgreementUpdate() when updateContactsAgreement != null:
-        return updateContactsAgreement(_that);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>({TResult? Function( _UserAppAgreementUpdate value)?  updateUserAgreement,TResult? Function( _ContactsAppAgreementUpdate value)?  updateContactsAgreement,}){
+final _that = this;
+switch (_that) {
+case _UserAppAgreementUpdate() when updateUserAgreement != null:
+return updateUserAgreement(_that);case _ContactsAppAgreementUpdate() when updateContactsAgreement != null:
+return updateContactsAgreement(_that);case _:
+  return null;
 
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function(AgreementStatus status)? updateUserAgreement,
-    TResult Function(AgreementStatus status)? updateContactsAgreement,
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _UserAppAgreementUpdate() when updateUserAgreement != null:
-        return updateUserAgreement(_that.status);
-      case _ContactsAppAgreementUpdate() when updateContactsAgreement != null:
-        return updateContactsAgreement(_that.status);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>({TResult Function( AgreementStatus status)?  updateUserAgreement,TResult Function( AgreementStatus status)?  updateContactsAgreement,required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _UserAppAgreementUpdate() when updateUserAgreement != null:
+return updateUserAgreement(_that.status);case _ContactsAppAgreementUpdate() when updateContactsAgreement != null:
+return updateContactsAgreement(_that.status);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function(AgreementStatus status) updateUserAgreement,
-    required TResult Function(AgreementStatus status) updateContactsAgreement,
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _UserAppAgreementUpdate():
-        return updateUserAgreement(_that.status);
-      case _ContactsAppAgreementUpdate():
-        return updateContactsAgreement(_that.status);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult when<TResult extends Object?>({required TResult Function( AgreementStatus status)  updateUserAgreement,required TResult Function( AgreementStatus status)  updateContactsAgreement,}) {final _that = this;
+switch (_that) {
+case _UserAppAgreementUpdate():
+return updateUserAgreement(_that.status);case _ContactsAppAgreementUpdate():
+return updateContactsAgreement(_that.status);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function(AgreementStatus status)? updateUserAgreement,
-    TResult? Function(AgreementStatus status)? updateContactsAgreement,
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _UserAppAgreementUpdate() when updateUserAgreement != null:
-        return updateUserAgreement(_that.status);
-      case _ContactsAppAgreementUpdate() when updateContactsAgreement != null:
-        return updateContactsAgreement(_that.status);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>({TResult? Function( AgreementStatus status)?  updateUserAgreement,TResult? Function( AgreementStatus status)?  updateContactsAgreement,}) {final _that = this;
+switch (_that) {
+case _UserAppAgreementUpdate() when updateUserAgreement != null:
+return updateUserAgreement(_that.status);case _ContactsAppAgreementUpdate() when updateContactsAgreement != null:
+return updateContactsAgreement(_that.status);case _:
+  return null;
+
+}
+}
+
 }
 
 /// @nodoc
+
 
 class _UserAppAgreementUpdate implements AppAgreementAccepted {
   const _UserAppAgreementUpdate(this.status);
+  
 
-  @override
-  final AgreementStatus status;
+@override final  AgreementStatus status;
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _UserAppAgreementUpdate &&
-            (identical(other.status, status) || other.status == status));
-  }
 
-  @override
-  int get hashCode => Object.hash(runtimeType, status);
 
-  @override
-  String toString() {
-    return 'AppAgreementAccepted.updateUserAgreement(status: $status)';
-  }
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _UserAppAgreementUpdate&&(identical(other.status, status) || other.status == status));
 }
 
+
+@override
+int get hashCode => Object.hash(runtimeType,status);
+
+@override
+String toString() {
+  return 'AppAgreementAccepted.updateUserAgreement(status: $status)';
+}
+
+
+}
+
+
+
+
 /// @nodoc
+
 
 class _ContactsAppAgreementUpdate implements AppAgreementAccepted {
   const _ContactsAppAgreementUpdate(this.status);
+  
 
-  @override
-  final AgreementStatus status;
+@override final  AgreementStatus status;
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _ContactsAppAgreementUpdate &&
-            (identical(other.status, status) || other.status == status));
-  }
 
-  @override
-  int get hashCode => Object.hash(runtimeType, status);
 
-  @override
-  String toString() {
-    return 'AppAgreementAccepted.updateContactsAgreement(status: $status)';
-  }
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _ContactsAppAgreementUpdate&&(identical(other.status, status) || other.status == status));
 }
+
+
+@override
+int get hashCode => Object.hash(runtimeType,status);
+
+@override
+String toString() {
+  return 'AppAgreementAccepted.updateContactsAgreement(status: $status)';
+}
+
+
+}
+
+
+
 
 /// @nodoc
 mixin _$AppState {
-  Session get session;
-  ThemeSettings get themeSettings;
-  ThemeMode get themeMode;
-  Locale get locale;
-  AgreementStatus get userAgreementStatus;
-  AgreementStatus get contactsAgreementStatus;
 
-  /// Create a copy of AppState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  $AppStateCopyWith<AppState> get copyWith =>
-      _$AppStateCopyWithImpl<AppState>(this as AppState, _$identity);
+ Session get session; ThemeSettings get themeSettings; ThemeMode get themeMode; Locale get locale; AgreementStatus get userAgreementStatus; AgreementStatus get contactsAgreementStatus;
+/// Create a copy of AppState
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$AppStateCopyWith<AppState> get copyWith => _$AppStateCopyWithImpl<AppState>(this as AppState, _$identity);
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is AppState &&
-            (identical(other.session, session) || other.session == session) &&
-            (identical(other.themeSettings, themeSettings) ||
-                other.themeSettings == themeSettings) &&
-            (identical(other.themeMode, themeMode) ||
-                other.themeMode == themeMode) &&
-            (identical(other.locale, locale) || other.locale == locale) &&
-            (identical(other.userAgreementStatus, userAgreementStatus) ||
-                other.userAgreementStatus == userAgreementStatus) &&
-            (identical(
-                    other.contactsAgreementStatus, contactsAgreementStatus) ||
-                other.contactsAgreementStatus == contactsAgreementStatus));
-  }
 
-  @override
-  int get hashCode => Object.hash(runtimeType, session, themeSettings,
-      themeMode, locale, userAgreementStatus, contactsAgreementStatus);
 
-  @override
-  String toString() {
-    return 'AppState(session: $session, themeSettings: $themeSettings, themeMode: $themeMode, locale: $locale, userAgreementStatus: $userAgreementStatus, contactsAgreementStatus: $contactsAgreementStatus)';
-  }
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is AppState&&(identical(other.session, session) || other.session == session)&&(identical(other.themeSettings, themeSettings) || other.themeSettings == themeSettings)&&(identical(other.themeMode, themeMode) || other.themeMode == themeMode)&&(identical(other.locale, locale) || other.locale == locale)&&(identical(other.userAgreementStatus, userAgreementStatus) || other.userAgreementStatus == userAgreementStatus)&&(identical(other.contactsAgreementStatus, contactsAgreementStatus) || other.contactsAgreementStatus == contactsAgreementStatus));
+}
+
+
+@override
+int get hashCode => Object.hash(runtimeType,session,themeSettings,themeMode,locale,userAgreementStatus,contactsAgreementStatus);
+
+@override
+String toString() {
+  return 'AppState(session: $session, themeSettings: $themeSettings, themeMode: $themeMode, locale: $locale, userAgreementStatus: $userAgreementStatus, contactsAgreementStatus: $contactsAgreementStatus)';
+}
+
+
 }
 
 /// @nodoc
-abstract mixin class $AppStateCopyWith<$Res> {
-  factory $AppStateCopyWith(AppState value, $Res Function(AppState) _then) =
-      _$AppStateCopyWithImpl;
-  @useResult
-  $Res call(
-      {Session session,
-      ThemeSettings themeSettings,
-      ThemeMode themeMode,
-      Locale locale,
-      AgreementStatus userAgreementStatus,
-      AgreementStatus contactsAgreementStatus});
+abstract mixin class $AppStateCopyWith<$Res>  {
+  factory $AppStateCopyWith(AppState value, $Res Function(AppState) _then) = _$AppStateCopyWithImpl;
+@useResult
+$Res call({
+ Session session, ThemeSettings themeSettings, ThemeMode themeMode, Locale locale, AgreementStatus userAgreementStatus, AgreementStatus contactsAgreementStatus
+});
 
-  $SessionCopyWith<$Res> get session;
-  $ThemeSettingsCopyWith<$Res> get themeSettings;
+
+
+
 }
-
 /// @nodoc
-class _$AppStateCopyWithImpl<$Res> implements $AppStateCopyWith<$Res> {
+class _$AppStateCopyWithImpl<$Res>
+    implements $AppStateCopyWith<$Res> {
   _$AppStateCopyWithImpl(this._self, this._then);
 
   final AppState _self;
   final $Res Function(AppState) _then;
 
-  /// Create a copy of AppState
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? session = null,
-    Object? themeSettings = null,
-    Object? themeMode = null,
-    Object? locale = null,
-    Object? userAgreementStatus = null,
-    Object? contactsAgreementStatus = null,
-  }) {
-    return _then(_self.copyWith(
-      session: null == session
-          ? _self.session
-          : session // ignore: cast_nullable_to_non_nullable
-              as Session,
-      themeSettings: null == themeSettings
-          ? _self.themeSettings
-          : themeSettings // ignore: cast_nullable_to_non_nullable
-              as ThemeSettings,
-      themeMode: null == themeMode
-          ? _self.themeMode
-          : themeMode // ignore: cast_nullable_to_non_nullable
-              as ThemeMode,
-      locale: null == locale
-          ? _self.locale
-          : locale // ignore: cast_nullable_to_non_nullable
-              as Locale,
-      userAgreementStatus: null == userAgreementStatus
-          ? _self.userAgreementStatus
-          : userAgreementStatus // ignore: cast_nullable_to_non_nullable
-              as AgreementStatus,
-      contactsAgreementStatus: null == contactsAgreementStatus
-          ? _self.contactsAgreementStatus
-          : contactsAgreementStatus // ignore: cast_nullable_to_non_nullable
-              as AgreementStatus,
-    ));
-  }
-
-  /// Create a copy of AppState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $SessionCopyWith<$Res> get session {
-    return $SessionCopyWith<$Res>(_self.session, (value) {
-      return _then(_self.copyWith(session: value));
-    });
-  }
-
-  /// Create a copy of AppState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $ThemeSettingsCopyWith<$Res> get themeSettings {
-    return $ThemeSettingsCopyWith<$Res>(_self.themeSettings, (value) {
-      return _then(_self.copyWith(themeSettings: value));
-    });
-  }
+/// Create a copy of AppState
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? session = null,Object? themeSettings = null,Object? themeMode = null,Object? locale = null,Object? userAgreementStatus = null,Object? contactsAgreementStatus = null,}) {
+  return _then(AppState(
+session: null == session ? _self.session : session // ignore: cast_nullable_to_non_nullable
+as Session,themeSettings: null == themeSettings ? _self.themeSettings : themeSettings // ignore: cast_nullable_to_non_nullable
+as ThemeSettings,themeMode: null == themeMode ? _self.themeMode : themeMode // ignore: cast_nullable_to_non_nullable
+as ThemeMode,locale: null == locale ? _self.locale : locale // ignore: cast_nullable_to_non_nullable
+as Locale,userAgreementStatus: null == userAgreementStatus ? _self.userAgreementStatus : userAgreementStatus // ignore: cast_nullable_to_non_nullable
+as AgreementStatus,contactsAgreementStatus: null == contactsAgreementStatus ? _self.contactsAgreementStatus : contactsAgreementStatus // ignore: cast_nullable_to_non_nullable
+as AgreementStatus,
+  ));
 }
+
+}
+
 
 /// Adds pattern-matching-related methods to [AppState].
 extension AppStatePatterns on AppState {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_AppState value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _AppState() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>({required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_AppState value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _AppState():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult map<TResult extends Object?>(){
+final _that = this;
+switch (_that) {
+case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_AppState value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _AppState() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(){
+final _that = this;
+switch (_that) {
+case _:
+  return null;
 
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(
-            Session session,
-            ThemeSettings themeSettings,
-            ThemeMode themeMode,
-            Locale locale,
-            AgreementStatus userAgreementStatus,
-            AgreementStatus contactsAgreementStatus)?
-        $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _AppState() when $default != null:
-        return $default(
-            _that.session,
-            _that.themeSettings,
-            _that.themeMode,
-            _that.locale,
-            _that.userAgreementStatus,
-            _that.contactsAgreementStatus);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>({required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(
-            Session session,
-            ThemeSettings themeSettings,
-            ThemeMode themeMode,
-            Locale locale,
-            AgreementStatus userAgreementStatus,
-            AgreementStatus contactsAgreementStatus)
-        $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _AppState():
-        return $default(
-            _that.session,
-            _that.themeSettings,
-            _that.themeMode,
-            _that.locale,
-            _that.userAgreementStatus,
-            _that.contactsAgreementStatus);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult when<TResult extends Object?>() {final _that = this;
+switch (_that) {
+case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(
-            Session session,
-            ThemeSettings themeSettings,
-            ThemeMode themeMode,
-            Locale locale,
-            AgreementStatus userAgreementStatus,
-            AgreementStatus contactsAgreementStatus)?
-        $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _AppState() when $default != null:
-        return $default(
-            _that.session,
-            _that.themeSettings,
-            _that.themeMode,
-            _that.locale,
-            _that.userAgreementStatus,
-            _that.contactsAgreementStatus);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>() {final _that = this;
+switch (_that) {
+case _:
+  return null;
+
+}
 }
 
-/// @nodoc
-
-class _AppState extends AppState {
-  const _AppState(
-      {this.session = const Session(),
-      required this.themeSettings,
-      required this.themeMode,
-      required this.locale,
-      required this.userAgreementStatus,
-      required this.contactsAgreementStatus})
-      : super._();
-
-  @override
-  @JsonKey()
-  final Session session;
-  @override
-  final ThemeSettings themeSettings;
-  @override
-  final ThemeMode themeMode;
-  @override
-  final Locale locale;
-  @override
-  final AgreementStatus userAgreementStatus;
-  @override
-  final AgreementStatus contactsAgreementStatus;
-
-  /// Create a copy of AppState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  _$AppStateCopyWith<_AppState> get copyWith =>
-      __$AppStateCopyWithImpl<_AppState>(this, _$identity);
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _AppState &&
-            (identical(other.session, session) || other.session == session) &&
-            (identical(other.themeSettings, themeSettings) ||
-                other.themeSettings == themeSettings) &&
-            (identical(other.themeMode, themeMode) ||
-                other.themeMode == themeMode) &&
-            (identical(other.locale, locale) || other.locale == locale) &&
-            (identical(other.userAgreementStatus, userAgreementStatus) ||
-                other.userAgreementStatus == userAgreementStatus) &&
-            (identical(
-                    other.contactsAgreementStatus, contactsAgreementStatus) ||
-                other.contactsAgreementStatus == contactsAgreementStatus));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, session, themeSettings,
-      themeMode, locale, userAgreementStatus, contactsAgreementStatus);
-
-  @override
-  String toString() {
-    return 'AppState(session: $session, themeSettings: $themeSettings, themeMode: $themeMode, locale: $locale, userAgreementStatus: $userAgreementStatus, contactsAgreementStatus: $contactsAgreementStatus)';
-  }
-}
-
-/// @nodoc
-abstract mixin class _$AppStateCopyWith<$Res>
-    implements $AppStateCopyWith<$Res> {
-  factory _$AppStateCopyWith(_AppState value, $Res Function(_AppState) _then) =
-      __$AppStateCopyWithImpl;
-  @override
-  @useResult
-  $Res call(
-      {Session session,
-      ThemeSettings themeSettings,
-      ThemeMode themeMode,
-      Locale locale,
-      AgreementStatus userAgreementStatus,
-      AgreementStatus contactsAgreementStatus});
-
-  @override
-  $SessionCopyWith<$Res> get session;
-  @override
-  $ThemeSettingsCopyWith<$Res> get themeSettings;
-}
-
-/// @nodoc
-class __$AppStateCopyWithImpl<$Res> implements _$AppStateCopyWith<$Res> {
-  __$AppStateCopyWithImpl(this._self, this._then);
-
-  final _AppState _self;
-  final $Res Function(_AppState) _then;
-
-  /// Create a copy of AppState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $Res call({
-    Object? session = null,
-    Object? themeSettings = null,
-    Object? themeMode = null,
-    Object? locale = null,
-    Object? userAgreementStatus = null,
-    Object? contactsAgreementStatus = null,
-  }) {
-    return _then(_AppState(
-      session: null == session
-          ? _self.session
-          : session // ignore: cast_nullable_to_non_nullable
-              as Session,
-      themeSettings: null == themeSettings
-          ? _self.themeSettings
-          : themeSettings // ignore: cast_nullable_to_non_nullable
-              as ThemeSettings,
-      themeMode: null == themeMode
-          ? _self.themeMode
-          : themeMode // ignore: cast_nullable_to_non_nullable
-              as ThemeMode,
-      locale: null == locale
-          ? _self.locale
-          : locale // ignore: cast_nullable_to_non_nullable
-              as Locale,
-      userAgreementStatus: null == userAgreementStatus
-          ? _self.userAgreementStatus
-          : userAgreementStatus // ignore: cast_nullable_to_non_nullable
-              as AgreementStatus,
-      contactsAgreementStatus: null == contactsAgreementStatus
-          ? _self.contactsAgreementStatus
-          : contactsAgreementStatus // ignore: cast_nullable_to_non_nullable
-              as AgreementStatus,
-    ));
-  }
-
-  /// Create a copy of AppState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $SessionCopyWith<$Res> get session {
-    return $SessionCopyWith<$Res>(_self.session, (value) {
-      return _then(_self.copyWith(session: value));
-    });
-  }
-
-  /// Create a copy of AppState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $ThemeSettingsCopyWith<$Res> get themeSettings {
-    return $ThemeSettingsCopyWith<$Res>(_self.themeSettings, (value) {
-      return _then(_self.copyWith(themeSettings: value));
-    });
-  }
 }
 
 // dart format on
