@@ -1,17 +1,29 @@
 part of 'app_bloc.dart';
 
 @freezed
-abstract class AppState with _$AppState {
-  const AppState._();
+class AppState with _$AppState {
 
-  const factory AppState({
-    @Default(Session()) Session session,
-    required ThemeSettings themeSettings,
-    required ThemeMode themeMode,
-    required Locale locale,
-    required AgreementStatus userAgreementStatus,
-    required AgreementStatus contactsAgreementStatus,
-  }) = _AppState;
+  const AppState({
+    this.session = const Session(),
+    required this.themeSettings,
+    required this.themeMode,
+    required this.locale,
+    required this.userAgreementStatus,
+    required this.contactsAgreementStatus,
+  });
+
+  @override
+  final Session session;
+  @override
+  final ThemeSettings themeSettings;
+  @override
+  final ThemeMode themeMode;
+  @override
+  final Locale locale;
+  @override
+  final AgreementStatus userAgreementStatus;
+  @override
+  final AgreementStatus contactsAgreementStatus;
 
   bool get isThemeModeSupported => themeSettings.lightColorSchemeConfig != themeSettings.darkColorSchemeConfig;
 
