@@ -1,6 +1,6 @@
 part of 'call_log_bloc.dart';
 
-abstract class CallLogEvent {
+sealed class CallLogEvent {
   const CallLogEvent();
 }
 
@@ -8,7 +8,8 @@ class CallLogStarted extends CallLogEvent {
   const CallLogStarted();
 }
 
-@Freezed(copyWith: false)
-abstract class CallLogEntryDeleted with _$CallLogEntryDeleted implements CallLogEvent {
-  const factory CallLogEntryDeleted(CallLogEntry callLogEntry) = _CallLogEntryDeleted;
+class CallLogEntryDeleted extends CallLogEvent {
+  const CallLogEntryDeleted(this.callLogEntry);
+
+  final CallLogEntry callLogEntry;
 }
