@@ -11,922 +11,190 @@ part of 'contact_bloc.dart';
 
 // dart format off
 T _$identity<T>(T value) => value;
-
-/// @nodoc
-mixin _$ContactAddedToFavorites {
-  ContactPhone get contactPhone;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is ContactAddedToFavorites &&
-            (identical(other.contactPhone, contactPhone) ||
-                other.contactPhone == contactPhone));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, contactPhone);
-
-  @override
-  String toString() {
-    return 'ContactAddedToFavorites(contactPhone: $contactPhone)';
-  }
-}
-
-/// Adds pattern-matching-related methods to [ContactAddedToFavorites].
-extension ContactAddedToFavoritesPatterns on ContactAddedToFavorites {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_ContactAddedToFavorites value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _ContactAddedToFavorites() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_ContactAddedToFavorites value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactAddedToFavorites():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_ContactAddedToFavorites value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactAddedToFavorites() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
-
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(ContactPhone contactPhone)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _ContactAddedToFavorites() when $default != null:
-        return $default(_that.contactPhone);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(ContactPhone contactPhone) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactAddedToFavorites():
-        return $default(_that.contactPhone);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(ContactPhone contactPhone)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactAddedToFavorites() when $default != null:
-        return $default(_that.contactPhone);
-      case _:
-        return null;
-    }
-  }
-}
-
-/// @nodoc
-
-class _ContactAddedToFavorites implements ContactAddedToFavorites {
-  const _ContactAddedToFavorites(this.contactPhone);
-
-  @override
-  final ContactPhone contactPhone;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _ContactAddedToFavorites &&
-            (identical(other.contactPhone, contactPhone) ||
-                other.contactPhone == contactPhone));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, contactPhone);
-
-  @override
-  String toString() {
-    return 'ContactAddedToFavorites(contactPhone: $contactPhone)';
-  }
-}
-
-/// @nodoc
-mixin _$ContactRemovedFromFavorites {
-  ContactPhone get contactPhone;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is ContactRemovedFromFavorites &&
-            (identical(other.contactPhone, contactPhone) ||
-                other.contactPhone == contactPhone));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, contactPhone);
-
-  @override
-  String toString() {
-    return 'ContactRemovedFromFavorites(contactPhone: $contactPhone)';
-  }
-}
-
-/// Adds pattern-matching-related methods to [ContactRemovedFromFavorites].
-extension ContactRemovedFromFavoritesPatterns on ContactRemovedFromFavorites {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_ContactRemovedFromFavorites value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _ContactRemovedFromFavorites() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_ContactRemovedFromFavorites value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactRemovedFromFavorites():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_ContactRemovedFromFavorites value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactRemovedFromFavorites() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
-
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(ContactPhone contactPhone)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _ContactRemovedFromFavorites() when $default != null:
-        return $default(_that.contactPhone);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(ContactPhone contactPhone) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactRemovedFromFavorites():
-        return $default(_that.contactPhone);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(ContactPhone contactPhone)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactRemovedFromFavorites() when $default != null:
-        return $default(_that.contactPhone);
-      case _:
-        return null;
-    }
-  }
-}
-
-/// @nodoc
-
-class _ContactRemovedFromFavorites implements ContactRemovedFromFavorites {
-  const _ContactRemovedFromFavorites(this.contactPhone);
-
-  @override
-  final ContactPhone contactPhone;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _ContactRemovedFromFavorites &&
-            (identical(other.contactPhone, contactPhone) ||
-                other.contactPhone == contactPhone));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, contactPhone);
-
-  @override
-  String toString() {
-    return 'ContactRemovedFromFavorites(contactPhone: $contactPhone)';
-  }
-}
-
-/// @nodoc
-mixin _$ContactEmailSend {
-  ContactEmail get contactEmail;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is ContactEmailSend &&
-            (identical(other.contactEmail, contactEmail) ||
-                other.contactEmail == contactEmail));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, contactEmail);
-
-  @override
-  String toString() {
-    return 'ContactEmailSend(contactEmail: $contactEmail)';
-  }
-}
-
-/// Adds pattern-matching-related methods to [ContactEmailSend].
-extension ContactEmailSendPatterns on ContactEmailSend {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_ContactEmailSend value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _ContactEmailSend() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_ContactEmailSend value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactEmailSend():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_ContactEmailSend value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactEmailSend() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
-
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(ContactEmail contactEmail)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _ContactEmailSend() when $default != null:
-        return $default(_that.contactEmail);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(ContactEmail contactEmail) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactEmailSend():
-        return $default(_that.contactEmail);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(ContactEmail contactEmail)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactEmailSend() when $default != null:
-        return $default(_that.contactEmail);
-      case _:
-        return null;
-    }
-  }
-}
-
-/// @nodoc
-
-class _ContactEmailSend implements ContactEmailSend {
-  const _ContactEmailSend(this.contactEmail);
-
-  @override
-  final ContactEmail contactEmail;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _ContactEmailSend &&
-            (identical(other.contactEmail, contactEmail) ||
-                other.contactEmail == contactEmail));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, contactEmail);
-
-  @override
-  String toString() {
-    return 'ContactEmailSend(contactEmail: $contactEmail)';
-  }
-}
-
 /// @nodoc
 mixin _$ContactState {
-  Contact? get contact;
-  bool get deleted;
 
-  /// Create a copy of ContactState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  $ContactStateCopyWith<ContactState> get copyWith =>
-      _$ContactStateCopyWithImpl<ContactState>(
-          this as ContactState, _$identity);
+ Contact? get contact; bool get deleted;
+/// Create a copy of ContactState
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$ContactStateCopyWith<ContactState> get copyWith => _$ContactStateCopyWithImpl<ContactState>(this as ContactState, _$identity);
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is ContactState &&
-            (identical(other.contact, contact) || other.contact == contact) &&
-            (identical(other.deleted, deleted) || other.deleted == deleted));
-  }
 
-  @override
-  int get hashCode => Object.hash(runtimeType, contact, deleted);
 
-  @override
-  String toString() {
-    return 'ContactState(contact: $contact, deleted: $deleted)';
-  }
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is ContactState&&(identical(other.contact, contact) || other.contact == contact)&&(identical(other.deleted, deleted) || other.deleted == deleted));
+}
+
+
+@override
+int get hashCode => Object.hash(runtimeType,contact,deleted);
+
+@override
+String toString() {
+  return 'ContactState(contact: $contact, deleted: $deleted)';
+}
+
+
 }
 
 /// @nodoc
-abstract mixin class $ContactStateCopyWith<$Res> {
-  factory $ContactStateCopyWith(
-          ContactState value, $Res Function(ContactState) _then) =
-      _$ContactStateCopyWithImpl;
-  @useResult
-  $Res call({Contact? contact, bool deleted});
-}
+abstract mixin class $ContactStateCopyWith<$Res>  {
+  factory $ContactStateCopyWith(ContactState value, $Res Function(ContactState) _then) = _$ContactStateCopyWithImpl;
+@useResult
+$Res call({
+ Contact? contact, bool deleted
+});
 
+
+
+
+}
 /// @nodoc
-class _$ContactStateCopyWithImpl<$Res> implements $ContactStateCopyWith<$Res> {
+class _$ContactStateCopyWithImpl<$Res>
+    implements $ContactStateCopyWith<$Res> {
   _$ContactStateCopyWithImpl(this._self, this._then);
 
   final ContactState _self;
   final $Res Function(ContactState) _then;
 
-  /// Create a copy of ContactState
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? contact = freezed,
-    Object? deleted = null,
-  }) {
-    return _then(_self.copyWith(
-      contact: freezed == contact
-          ? _self.contact
-          : contact // ignore: cast_nullable_to_non_nullable
-              as Contact?,
-      deleted: null == deleted
-          ? _self.deleted
-          : deleted // ignore: cast_nullable_to_non_nullable
-              as bool,
-    ));
-  }
+/// Create a copy of ContactState
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? contact = freezed,Object? deleted = null,}) {
+  return _then(ContactState(
+contact: freezed == contact ? _self.contact : contact // ignore: cast_nullable_to_non_nullable
+as Contact?,deleted: null == deleted ? _self.deleted : deleted // ignore: cast_nullable_to_non_nullable
+as bool,
+  ));
 }
+
+}
+
 
 /// Adds pattern-matching-related methods to [ContactState].
 extension ContactStatePatterns on ContactState {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_ContactState value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _ContactState() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>({required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_ContactState value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactState():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult map<TResult extends Object?>(){
+final _that = this;
+switch (_that) {
+case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_ContactState value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactState() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(){
+final _that = this;
+switch (_that) {
+case _:
+  return null;
 
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(Contact? contact, bool deleted)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _ContactState() when $default != null:
-        return $default(_that.contact, _that.deleted);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>({required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(Contact? contact, bool deleted) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactState():
-        return $default(_that.contact, _that.deleted);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult when<TResult extends Object?>() {final _that = this;
+switch (_that) {
+case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(Contact? contact, bool deleted)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactState() when $default != null:
-        return $default(_that.contact, _that.deleted);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>() {final _that = this;
+switch (_that) {
+case _:
+  return null;
+
+}
 }
 
-/// @nodoc
-
-class _ContactState implements ContactState {
-  const _ContactState({this.contact, this.deleted = false});
-
-  @override
-  final Contact? contact;
-  @override
-  @JsonKey()
-  final bool deleted;
-
-  /// Create a copy of ContactState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  _$ContactStateCopyWith<_ContactState> get copyWith =>
-      __$ContactStateCopyWithImpl<_ContactState>(this, _$identity);
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _ContactState &&
-            (identical(other.contact, contact) || other.contact == contact) &&
-            (identical(other.deleted, deleted) || other.deleted == deleted));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, contact, deleted);
-
-  @override
-  String toString() {
-    return 'ContactState(contact: $contact, deleted: $deleted)';
-  }
-}
-
-/// @nodoc
-abstract mixin class _$ContactStateCopyWith<$Res>
-    implements $ContactStateCopyWith<$Res> {
-  factory _$ContactStateCopyWith(
-          _ContactState value, $Res Function(_ContactState) _then) =
-      __$ContactStateCopyWithImpl;
-  @override
-  @useResult
-  $Res call({Contact? contact, bool deleted});
-}
-
-/// @nodoc
-class __$ContactStateCopyWithImpl<$Res>
-    implements _$ContactStateCopyWith<$Res> {
-  __$ContactStateCopyWithImpl(this._self, this._then);
-
-  final _ContactState _self;
-  final $Res Function(_ContactState) _then;
-
-  /// Create a copy of ContactState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $Res call({
-    Object? contact = freezed,
-    Object? deleted = null,
-  }) {
-    return _then(_ContactState(
-      contact: freezed == contact
-          ? _self.contact
-          : contact // ignore: cast_nullable_to_non_nullable
-              as Contact?,
-      deleted: null == deleted
-          ? _self.deleted
-          : deleted // ignore: cast_nullable_to_non_nullable
-              as bool,
-    ));
-  }
 }
 
 // dart format on
