@@ -1,6 +1,6 @@
 part of 'contact_bloc.dart';
 
-abstract class ContactEvent {
+sealed class ContactEvent {
   const ContactEvent();
 }
 
@@ -8,17 +8,20 @@ class ContactStarted extends ContactEvent {
   const ContactStarted();
 }
 
-@Freezed(copyWith: false)
-abstract class ContactAddedToFavorites with _$ContactAddedToFavorites implements ContactEvent {
-  const factory ContactAddedToFavorites(ContactPhone contactPhone) = _ContactAddedToFavorites;
+class ContactAddedToFavorites extends ContactEvent {
+  const ContactAddedToFavorites(this.contactPhone);
+
+  final ContactPhone contactPhone;
 }
 
-@Freezed(copyWith: false)
-abstract class ContactRemovedFromFavorites with _$ContactRemovedFromFavorites implements ContactEvent {
-  const factory ContactRemovedFromFavorites(ContactPhone contactPhone) = _ContactRemovedFromFavorites;
+class ContactRemovedFromFavorites extends ContactEvent {
+  const ContactRemovedFromFavorites(this.contactPhone);
+
+  final ContactPhone contactPhone;
 }
 
-@Freezed(copyWith: false)
-abstract class ContactEmailSend with _$ContactEmailSend implements ContactEvent {
-  const factory ContactEmailSend(ContactEmail contactEmail) = _ContactEmailSend;
+class ContactEmailSend extends ContactEvent {
+  const ContactEmailSend(this.contactEmail);
+
+  final ContactEmail contactEmail;
 }
