@@ -11,659 +11,47 @@ part of 'contacts_bloc.dart';
 
 // dart format off
 T _$identity<T>(T value) => value;
-
-/// @nodoc
-mixin _$ContactsSourceTypeChanged {
-  ContactSourceType get sourceType;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is ContactsSourceTypeChanged &&
-            (identical(other.sourceType, sourceType) ||
-                other.sourceType == sourceType));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, sourceType);
-
-  @override
-  String toString() {
-    return 'ContactsSourceTypeChanged(sourceType: $sourceType)';
-  }
-}
-
-/// Adds pattern-matching-related methods to [ContactsSourceTypeChanged].
-extension ContactsSourceTypeChangedPatterns on ContactsSourceTypeChanged {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_ContactsSourceTypeChanged value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSourceTypeChanged() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_ContactsSourceTypeChanged value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSourceTypeChanged():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_ContactsSourceTypeChanged value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSourceTypeChanged() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
-
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(ContactSourceType sourceType)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSourceTypeChanged() when $default != null:
-        return $default(_that.sourceType);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(ContactSourceType sourceType) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSourceTypeChanged():
-        return $default(_that.sourceType);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(ContactSourceType sourceType)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSourceTypeChanged() when $default != null:
-        return $default(_that.sourceType);
-      case _:
-        return null;
-    }
-  }
-}
-
-/// @nodoc
-
-class _ContactsSourceTypeChanged implements ContactsSourceTypeChanged {
-  const _ContactsSourceTypeChanged(this.sourceType);
-
-  @override
-  final ContactSourceType sourceType;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _ContactsSourceTypeChanged &&
-            (identical(other.sourceType, sourceType) ||
-                other.sourceType == sourceType));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, sourceType);
-
-  @override
-  String toString() {
-    return 'ContactsSourceTypeChanged(sourceType: $sourceType)';
-  }
-}
-
-/// @nodoc
-mixin _$ContactsSearchChanged {
-  String get search;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is ContactsSearchChanged &&
-            (identical(other.search, search) || other.search == search));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, search);
-
-  @override
-  String toString() {
-    return 'ContactsSearchChanged(search: $search)';
-  }
-}
-
-/// Adds pattern-matching-related methods to [ContactsSearchChanged].
-extension ContactsSearchChangedPatterns on ContactsSearchChanged {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_ContactsSearchChanged value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSearchChanged() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_ContactsSearchChanged value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSearchChanged():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_ContactsSearchChanged value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSearchChanged() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
-
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(String search)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSearchChanged() when $default != null:
-        return $default(_that.search);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(String search) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSearchChanged():
-        return $default(_that.search);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(String search)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSearchChanged() when $default != null:
-        return $default(_that.search);
-      case _:
-        return null;
-    }
-  }
-}
-
-/// @nodoc
-
-class _ContactsSearchChanged implements ContactsSearchChanged {
-  const _ContactsSearchChanged(this.search);
-
-  @override
-  final String search;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _ContactsSearchChanged &&
-            (identical(other.search, search) || other.search == search));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, search);
-
-  @override
-  String toString() {
-    return 'ContactsSearchChanged(search: $search)';
-  }
-}
-
-/// @nodoc
-mixin _$ContactsSearchSubmitted {
-  String get search;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is ContactsSearchSubmitted &&
-            (identical(other.search, search) || other.search == search));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, search);
-
-  @override
-  String toString() {
-    return 'ContactsSearchSubmitted(search: $search)';
-  }
-}
-
-/// Adds pattern-matching-related methods to [ContactsSearchSubmitted].
-extension ContactsSearchSubmittedPatterns on ContactsSearchSubmitted {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_ContactsSearchSubmitted value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSearchSubmitted() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_ContactsSearchSubmitted value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSearchSubmitted():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_ContactsSearchSubmitted value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSearchSubmitted() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
-
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(String search)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSearchSubmitted() when $default != null:
-        return $default(_that.search);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(String search) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSearchSubmitted():
-        return $default(_that.search);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(String search)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsSearchSubmitted() when $default != null:
-        return $default(_that.search);
-      case _:
-        return null;
-    }
-  }
-}
-
-/// @nodoc
-
-class _ContactsSearchSubmitted implements ContactsSearchSubmitted {
-  const _ContactsSearchSubmitted(this.search);
-
-  @override
-  final String search;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _ContactsSearchSubmitted &&
-            (identical(other.search, search) || other.search == search));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, search);
-
-  @override
-  String toString() {
-    return 'ContactsSearchSubmitted(search: $search)';
-  }
-}
-
 /// @nodoc
 mixin _$ContactsState {
-  String get search;
-  ContactSourceType get sourceType;
 
-  /// Create a copy of ContactsState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  $ContactsStateCopyWith<ContactsState> get copyWith =>
-      _$ContactsStateCopyWithImpl<ContactsState>(
-          this as ContactsState, _$identity);
+ String get search; ContactSourceType get sourceType;
+/// Create a copy of ContactsState
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$ContactsStateCopyWith<ContactsState> get copyWith => _$ContactsStateCopyWithImpl<ContactsState>(this as ContactsState, _$identity);
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is ContactsState &&
-            (identical(other.search, search) || other.search == search) &&
-            (identical(other.sourceType, sourceType) ||
-                other.sourceType == sourceType));
-  }
 
-  @override
-  int get hashCode => Object.hash(runtimeType, search, sourceType);
 
-  @override
-  String toString() {
-    return 'ContactsState(search: $search, sourceType: $sourceType)';
-  }
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is ContactsState&&(identical(other.search, search) || other.search == search)&&(identical(other.sourceType, sourceType) || other.sourceType == sourceType));
+}
+
+
+@override
+int get hashCode => Object.hash(runtimeType,search,sourceType);
+
+@override
+String toString() {
+  return 'ContactsState(search: $search, sourceType: $sourceType)';
+}
+
+
 }
 
 /// @nodoc
-abstract mixin class $ContactsStateCopyWith<$Res> {
-  factory $ContactsStateCopyWith(
-          ContactsState value, $Res Function(ContactsState) _then) =
-      _$ContactsStateCopyWithImpl;
-  @useResult
-  $Res call({String search, ContactSourceType sourceType});
-}
+abstract mixin class $ContactsStateCopyWith<$Res>  {
+  factory $ContactsStateCopyWith(ContactsState value, $Res Function(ContactsState) _then) = _$ContactsStateCopyWithImpl;
+@useResult
+$Res call({
+ String search, ContactSourceType sourceType
+});
 
+
+
+
+}
 /// @nodoc
 class _$ContactsStateCopyWithImpl<$Res>
     implements $ContactsStateCopyWith<$Res> {
@@ -672,260 +60,141 @@ class _$ContactsStateCopyWithImpl<$Res>
   final ContactsState _self;
   final $Res Function(ContactsState) _then;
 
-  /// Create a copy of ContactsState
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? search = null,
-    Object? sourceType = null,
-  }) {
-    return _then(_self.copyWith(
-      search: null == search
-          ? _self.search
-          : search // ignore: cast_nullable_to_non_nullable
-              as String,
-      sourceType: null == sourceType
-          ? _self.sourceType
-          : sourceType // ignore: cast_nullable_to_non_nullable
-              as ContactSourceType,
-    ));
-  }
+/// Create a copy of ContactsState
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? search = null,Object? sourceType = null,}) {
+  return _then(ContactsState(
+search: null == search ? _self.search : search // ignore: cast_nullable_to_non_nullable
+as String,sourceType: null == sourceType ? _self.sourceType : sourceType // ignore: cast_nullable_to_non_nullable
+as ContactSourceType,
+  ));
 }
+
+}
+
 
 /// Adds pattern-matching-related methods to [ContactsState].
 extension ContactsStatePatterns on ContactsState {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_ContactsState value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsState() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>({required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_ContactsState value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsState():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult map<TResult extends Object?>(){
+final _that = this;
+switch (_that) {
+case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_ContactsState value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsState() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(){
+final _that = this;
+switch (_that) {
+case _:
+  return null;
 
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(String search, ContactSourceType sourceType)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsState() when $default != null:
-        return $default(_that.search, _that.sourceType);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>({required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(String search, ContactSourceType sourceType) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsState():
-        return $default(_that.search, _that.sourceType);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult when<TResult extends Object?>() {final _that = this;
+switch (_that) {
+case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(String search, ContactSourceType sourceType)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _ContactsState() when $default != null:
-        return $default(_that.search, _that.sourceType);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>() {final _that = this;
+switch (_that) {
+case _:
+  return null;
+
+}
 }
 
-/// @nodoc
-
-class _ContactsState implements ContactsState {
-  const _ContactsState({this.search = '', required this.sourceType});
-
-  @override
-  @JsonKey()
-  final String search;
-  @override
-  final ContactSourceType sourceType;
-
-  /// Create a copy of ContactsState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  _$ContactsStateCopyWith<_ContactsState> get copyWith =>
-      __$ContactsStateCopyWithImpl<_ContactsState>(this, _$identity);
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _ContactsState &&
-            (identical(other.search, search) || other.search == search) &&
-            (identical(other.sourceType, sourceType) ||
-                other.sourceType == sourceType));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, search, sourceType);
-
-  @override
-  String toString() {
-    return 'ContactsState(search: $search, sourceType: $sourceType)';
-  }
-}
-
-/// @nodoc
-abstract mixin class _$ContactsStateCopyWith<$Res>
-    implements $ContactsStateCopyWith<$Res> {
-  factory _$ContactsStateCopyWith(
-          _ContactsState value, $Res Function(_ContactsState) _then) =
-      __$ContactsStateCopyWithImpl;
-  @override
-  @useResult
-  $Res call({String search, ContactSourceType sourceType});
-}
-
-/// @nodoc
-class __$ContactsStateCopyWithImpl<$Res>
-    implements _$ContactsStateCopyWith<$Res> {
-  __$ContactsStateCopyWithImpl(this._self, this._then);
-
-  final _ContactsState _self;
-  final $Res Function(_ContactsState) _then;
-
-  /// Create a copy of ContactsState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $Res call({
-    Object? search = null,
-    Object? sourceType = null,
-  }) {
-    return _then(_ContactsState(
-      search: null == search
-          ? _self.search
-          : search // ignore: cast_nullable_to_non_nullable
-              as String,
-      sourceType: null == sourceType
-          ? _self.sourceType
-          : sourceType // ignore: cast_nullable_to_non_nullable
-              as ContactSourceType,
-    ));
-  }
 }
 
 // dart format on
