@@ -5,16 +5,25 @@ enum EmbeddedIntents {
 }
 
 @freezed
-abstract class EmbeddedState with _$EmbeddedState {
-  const EmbeddedState._();
+class EmbeddedState with _$EmbeddedState {
+  const EmbeddedState({
+    this.payload = const {},
+    this.currentUrl = '',
+    this.canGoBack = false,
+    this.payloadReady = false,
+    this.intent,
+  });
 
-  const factory EmbeddedState({
-    @Default({}) Map<String, dynamic> payload,
-    @Default('') String currentUrl,
-    @Default(false) bool canGoBack,
-    @Default(false) bool payloadReady,
-    EmbeddedIntents? intent,
-  }) = _Initial;
+  @override
+  final Map<String, dynamic> payload;
+  @override
+  final String currentUrl;
+  @override
+  final bool canGoBack;
+  @override
+  final bool payloadReady;
+  @override
+  final EmbeddedIntents? intent;
 
   bool get isReadyToInjectedScript => payloadReady;
 
