@@ -11,859 +11,47 @@ part of 'favorites_bloc.dart';
 
 // dart format off
 T _$identity<T>(T value) => value;
-
-/// @nodoc
-mixin _$FavoritesStarted {
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType && other is FavoritesStarted);
-  }
-
-  @override
-  int get hashCode => runtimeType.hashCode;
-
-  @override
-  String toString() {
-    return 'FavoritesStarted()';
-  }
-}
-
-/// Adds pattern-matching-related methods to [FavoritesStarted].
-extension FavoritesStartedPatterns on FavoritesStarted {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_FavoritesStarted value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesStarted() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_FavoritesStarted value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesStarted():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_FavoritesStarted value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesStarted() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
-
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function()? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesStarted() when $default != null:
-        return $default();
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function() $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesStarted():
-        return $default();
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function()? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesStarted() when $default != null:
-        return $default();
-      case _:
-        return null;
-    }
-  }
-}
-
-/// @nodoc
-
-class _FavoritesStarted implements FavoritesStarted {
-  const _FavoritesStarted();
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType && other is _FavoritesStarted);
-  }
-
-  @override
-  int get hashCode => runtimeType.hashCode;
-
-  @override
-  String toString() {
-    return 'FavoritesStarted()';
-  }
-}
-
-/// @nodoc
-mixin _$FavoritesAddedByContactPhoneId {
-  int get contactPhoneId;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is FavoritesAddedByContactPhoneId &&
-            (identical(other.contactPhoneId, contactPhoneId) ||
-                other.contactPhoneId == contactPhoneId));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, contactPhoneId);
-
-  @override
-  String toString() {
-    return 'FavoritesAddedByContactPhoneId(contactPhoneId: $contactPhoneId)';
-  }
-}
-
-/// Adds pattern-matching-related methods to [FavoritesAddedByContactPhoneId].
-extension FavoritesAddedByContactPhoneIdPatterns
-    on FavoritesAddedByContactPhoneId {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_FavoritesAddedByContactPhoneId value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesAddedByContactPhoneId() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_FavoritesAddedByContactPhoneId value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesAddedByContactPhoneId():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_FavoritesAddedByContactPhoneId value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesAddedByContactPhoneId() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
-
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(int contactPhoneId)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesAddedByContactPhoneId() when $default != null:
-        return $default(_that.contactPhoneId);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(int contactPhoneId) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesAddedByContactPhoneId():
-        return $default(_that.contactPhoneId);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(int contactPhoneId)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesAddedByContactPhoneId() when $default != null:
-        return $default(_that.contactPhoneId);
-      case _:
-        return null;
-    }
-  }
-}
-
-/// @nodoc
-
-class _FavoritesAddedByContactPhoneId
-    implements FavoritesAddedByContactPhoneId {
-  const _FavoritesAddedByContactPhoneId({required this.contactPhoneId});
-
-  @override
-  final int contactPhoneId;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _FavoritesAddedByContactPhoneId &&
-            (identical(other.contactPhoneId, contactPhoneId) ||
-                other.contactPhoneId == contactPhoneId));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, contactPhoneId);
-
-  @override
-  String toString() {
-    return 'FavoritesAddedByContactPhoneId(contactPhoneId: $contactPhoneId)';
-  }
-}
-
-/// @nodoc
-mixin _$FavoritesRemovedByContactPhoneId {
-  int get contactPhoneId;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is FavoritesRemovedByContactPhoneId &&
-            (identical(other.contactPhoneId, contactPhoneId) ||
-                other.contactPhoneId == contactPhoneId));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, contactPhoneId);
-
-  @override
-  String toString() {
-    return 'FavoritesRemovedByContactPhoneId(contactPhoneId: $contactPhoneId)';
-  }
-}
-
-/// Adds pattern-matching-related methods to [FavoritesRemovedByContactPhoneId].
-extension FavoritesRemovedByContactPhoneIdPatterns
-    on FavoritesRemovedByContactPhoneId {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_FavoritesRemovedByContactPhoneId value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesRemovedByContactPhoneId() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_FavoritesRemovedByContactPhoneId value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesRemovedByContactPhoneId():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_FavoritesRemovedByContactPhoneId value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesRemovedByContactPhoneId() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
-
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(int contactPhoneId)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesRemovedByContactPhoneId() when $default != null:
-        return $default(_that.contactPhoneId);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(int contactPhoneId) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesRemovedByContactPhoneId():
-        return $default(_that.contactPhoneId);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(int contactPhoneId)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesRemovedByContactPhoneId() when $default != null:
-        return $default(_that.contactPhoneId);
-      case _:
-        return null;
-    }
-  }
-}
-
-/// @nodoc
-
-class _FavoritesRemovedByContactPhoneId
-    implements FavoritesRemovedByContactPhoneId {
-  const _FavoritesRemovedByContactPhoneId({required this.contactPhoneId});
-
-  @override
-  final int contactPhoneId;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _FavoritesRemovedByContactPhoneId &&
-            (identical(other.contactPhoneId, contactPhoneId) ||
-                other.contactPhoneId == contactPhoneId));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, contactPhoneId);
-
-  @override
-  String toString() {
-    return 'FavoritesRemovedByContactPhoneId(contactPhoneId: $contactPhoneId)';
-  }
-}
-
-/// @nodoc
-mixin _$FavoritesRemoved {
-  Favorite get favorite;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is FavoritesRemoved &&
-            (identical(other.favorite, favorite) ||
-                other.favorite == favorite));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, favorite);
-
-  @override
-  String toString() {
-    return 'FavoritesRemoved(favorite: $favorite)';
-  }
-}
-
-/// Adds pattern-matching-related methods to [FavoritesRemoved].
-extension FavoritesRemovedPatterns on FavoritesRemoved {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_FavoritesRemoved value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesRemoved() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_FavoritesRemoved value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesRemoved():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_FavoritesRemoved value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesRemoved() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
-
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(Favorite favorite)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesRemoved() when $default != null:
-        return $default(_that.favorite);
-      case _:
-        return orElse();
-    }
-  }
-
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(Favorite favorite) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesRemoved():
-        return $default(_that.favorite);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
-
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
-
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(Favorite favorite)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesRemoved() when $default != null:
-        return $default(_that.favorite);
-      case _:
-        return null;
-    }
-  }
-}
-
-/// @nodoc
-
-class _FavoritesRemoved implements FavoritesRemoved {
-  const _FavoritesRemoved({required this.favorite});
-
-  @override
-  final Favorite favorite;
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _FavoritesRemoved &&
-            (identical(other.favorite, favorite) ||
-                other.favorite == favorite));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, favorite);
-
-  @override
-  String toString() {
-    return 'FavoritesRemoved(favorite: $favorite)';
-  }
-}
-
 /// @nodoc
 mixin _$FavoritesState {
-  List<Favorite>? get favorites;
 
-  /// Create a copy of FavoritesState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  $FavoritesStateCopyWith<FavoritesState> get copyWith =>
-      _$FavoritesStateCopyWithImpl<FavoritesState>(
-          this as FavoritesState, _$identity);
+ List<Favorite>? get favorites;
+/// Create a copy of FavoritesState
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$FavoritesStateCopyWith<FavoritesState> get copyWith => _$FavoritesStateCopyWithImpl<FavoritesState>(this as FavoritesState, _$identity);
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is FavoritesState &&
-            const DeepCollectionEquality().equals(other.favorites, favorites));
-  }
 
-  @override
-  int get hashCode =>
-      Object.hash(runtimeType, const DeepCollectionEquality().hash(favorites));
 
-  @override
-  String toString() {
-    return 'FavoritesState(favorites: $favorites)';
-  }
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is FavoritesState&&const DeepCollectionEquality().equals(other.favorites, favorites));
+}
+
+
+@override
+int get hashCode => Object.hash(runtimeType,const DeepCollectionEquality().hash(favorites));
+
+@override
+String toString() {
+  return 'FavoritesState(favorites: $favorites)';
+}
+
+
 }
 
 /// @nodoc
-abstract mixin class $FavoritesStateCopyWith<$Res> {
-  factory $FavoritesStateCopyWith(
-          FavoritesState value, $Res Function(FavoritesState) _then) =
-      _$FavoritesStateCopyWithImpl;
-  @useResult
-  $Res call({List<Favorite>? favorites});
-}
+abstract mixin class $FavoritesStateCopyWith<$Res>  {
+  factory $FavoritesStateCopyWith(FavoritesState value, $Res Function(FavoritesState) _then) = _$FavoritesStateCopyWithImpl;
+@useResult
+$Res call({
+ List<Favorite>? favorites
+});
 
+
+
+
+}
 /// @nodoc
 class _$FavoritesStateCopyWithImpl<$Res>
     implements $FavoritesStateCopyWith<$Res> {
@@ -872,255 +60,140 @@ class _$FavoritesStateCopyWithImpl<$Res>
   final FavoritesState _self;
   final $Res Function(FavoritesState) _then;
 
-  /// Create a copy of FavoritesState
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? favorites = freezed,
-  }) {
-    return _then(_self.copyWith(
-      favorites: freezed == favorites
-          ? _self.favorites
-          : favorites // ignore: cast_nullable_to_non_nullable
-              as List<Favorite>?,
-    ));
-  }
+/// Create a copy of FavoritesState
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? favorites = freezed,}) {
+  return _then(FavoritesState(
+favorites: freezed == favorites ? _self.favorites : favorites // ignore: cast_nullable_to_non_nullable
+as List<Favorite>?,
+  ));
 }
+
+}
+
 
 /// Adds pattern-matching-related methods to [FavoritesState].
 extension FavoritesStatePatterns on FavoritesState {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_FavoritesState value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesState() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>({required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_FavoritesState value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesState():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult map<TResult extends Object?>(){
+final _that = this;
+switch (_that) {
+case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_FavoritesState value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesState() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(){
+final _that = this;
+switch (_that) {
+case _:
+  return null;
 
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(List<Favorite>? favorites)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesState() when $default != null:
-        return $default(_that.favorites);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>({required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(List<Favorite>? favorites) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesState():
-        return $default(_that.favorites);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult when<TResult extends Object?>() {final _that = this;
+switch (_that) {
+case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(List<Favorite>? favorites)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _FavoritesState() when $default != null:
-        return $default(_that.favorites);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>() {final _that = this;
+switch (_that) {
+case _:
+  return null;
+
+}
 }
 
-/// @nodoc
-
-class _FavoritesState implements FavoritesState {
-  const _FavoritesState({final List<Favorite>? favorites})
-      : _favorites = favorites;
-
-  final List<Favorite>? _favorites;
-  @override
-  List<Favorite>? get favorites {
-    final value = _favorites;
-    if (value == null) return null;
-    if (_favorites is EqualUnmodifiableListView) return _favorites;
-    // ignore: implicit_dynamic_type
-    return EqualUnmodifiableListView(value);
-  }
-
-  /// Create a copy of FavoritesState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  _$FavoritesStateCopyWith<_FavoritesState> get copyWith =>
-      __$FavoritesStateCopyWithImpl<_FavoritesState>(this, _$identity);
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _FavoritesState &&
-            const DeepCollectionEquality()
-                .equals(other._favorites, _favorites));
-  }
-
-  @override
-  int get hashCode =>
-      Object.hash(runtimeType, const DeepCollectionEquality().hash(_favorites));
-
-  @override
-  String toString() {
-    return 'FavoritesState(favorites: $favorites)';
-  }
-}
-
-/// @nodoc
-abstract mixin class _$FavoritesStateCopyWith<$Res>
-    implements $FavoritesStateCopyWith<$Res> {
-  factory _$FavoritesStateCopyWith(
-          _FavoritesState value, $Res Function(_FavoritesState) _then) =
-      __$FavoritesStateCopyWithImpl;
-  @override
-  @useResult
-  $Res call({List<Favorite>? favorites});
-}
-
-/// @nodoc
-class __$FavoritesStateCopyWithImpl<$Res>
-    implements _$FavoritesStateCopyWith<$Res> {
-  __$FavoritesStateCopyWithImpl(this._self, this._then);
-
-  final _FavoritesState _self;
-  final $Res Function(_FavoritesState) _then;
-
-  /// Create a copy of FavoritesState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $Res call({
-    Object? favorites = freezed,
-  }) {
-    return _then(_FavoritesState(
-      favorites: freezed == favorites
-          ? _self._favorites
-          : favorites // ignore: cast_nullable_to_non_nullable
-              as List<Favorite>?,
-    ));
-  }
 }
 
 // dart format on
