@@ -1,8 +1,11 @@
 part of 'favorites_bloc.dart';
 
 @freezed
-abstract class FavoritesState with _$FavoritesState {
-  const factory FavoritesState({
-    List<Favorite>? favorites,
-  }) = _FavoritesState;
+class FavoritesState with _$FavoritesState {
+  const FavoritesState({
+    this.favorites,
+  });
+
+  @override
+  final List<Favorite>? favorites;
 }
