@@ -1,8 +1,11 @@
 part of 'keypad_cubit.dart';
 
 @freezed
-abstract class KeypadState with _$KeypadState {
-  const factory KeypadState({
-    Contact? contact,
-  }) = _KeypadState;
+class KeypadState with _$KeypadState {
+  const KeypadState({
+    this.contact,
+  });
+
+  @override
+  final Contact? contact;
 }
