@@ -1,6 +1,6 @@
 part of 'messaging_bloc.dart';
 
-abstract class MessagingEvent {
+sealed class MessagingEvent {
   const MessagingEvent();
 }
 
