@@ -1,10 +1,11 @@
 part of 'orientations_bloc.dart';
 
-abstract class OrientationsEvent {
+sealed class OrientationsEvent {
   const OrientationsEvent();
 }
 
-@Freezed(copyWith: false)
-abstract class OrientationsChanged with _$OrientationsChanged implements OrientationsEvent {
-  const factory OrientationsChanged(PreferredOrientation orientation) = _OrientationsChanged;
+class OrientationsChanged extends OrientationsEvent {
+  const OrientationsChanged(this.orientation);
+
+  final PreferredOrientation orientation;
 }
