@@ -1,6 +1,11 @@
 part of 'orientations_bloc.dart';
 
 @freezed
-abstract class OrientationsState with _$OrientationsState {
-  const factory OrientationsState([PreferredOrientation? lastOrientation]) = _OrientationsState;
+class OrientationsState with _$OrientationsState {
+  const OrientationsState({
+    this.lastOrientation,
+  });
+
+  @override
+  final PreferredOrientation? lastOrientation;
 }
