@@ -1,11 +1,17 @@
 part of 'push_tokens_bloc.dart';
 
 @freezed
-abstract class PushTokensState with _$PushTokensState {
-  const factory PushTokensState({
-    String? pushToken,
-    String? errorMessage,
-  }) = _PushTokensState;
+class PushTokensState with _$PushTokensState {
+  const PushTokensState({
+    this.pushToken,
+    this.errorMessage,
+  });
+
+  @override
+  final String? errorMessage;
+
+  @override
+  final String? pushToken;
 
   factory PushTokensState.initial() => const PushTokensState();
 
