@@ -1,6 +1,6 @@
 part of 'recents_bloc.dart';
 
-abstract class RecentsEvent {
+sealed class RecentsEvent {
   const RecentsEvent();
 }
 
@@ -8,12 +8,14 @@ class RecentsStarted extends RecentsEvent {
   const RecentsStarted();
 }
 
-@Freezed(copyWith: false)
-abstract class RecentsFiltered with _$RecentsFiltered implements RecentsEvent {
-  const factory RecentsFiltered(RecentsVisibilityFilter filter) = _RecentsFiltered;
+class RecentsFiltered extends RecentsEvent {
+  const RecentsFiltered(this.filter);
+
+  final RecentsVisibilityFilter filter;
 }
 
-@Freezed(copyWith: false)
-abstract class RecentsDeleted with _$RecentsDeleted implements RecentsEvent {
-  const factory RecentsDeleted(Recent recent) = _RecentsDeleted;
+class RecentsDeleted extends RecentsEvent {
+  const RecentsDeleted(this.recent);
+
+  final Recent recent;
 }
