@@ -1,8 +1,11 @@
 part of 'session_status_cubit.dart';
 
 @freezed
-abstract class SessionStatusState with _$SessionStatusState {
-  const factory SessionStatusState.initial({
-    @Default(SessionStatus.inProgress) SessionStatus status,
-  }) = _Initial;
+class SessionStatusState with _$SessionStatusState {
+  const SessionStatusState({
+    this.status = SessionStatus.inProgress,
+  });
+
+  @override
+  final SessionStatus status;
 }
