@@ -11,49 +11,47 @@ part of 'settings_bloc.dart';
 
 // dart format off
 T _$identity<T>(T value) => value;
-
 /// @nodoc
 mixin _$SettingsState {
-  bool get progress;
-  int get unreadVoicemailCount;
 
-  /// Create a copy of SettingsState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  $SettingsStateCopyWith<SettingsState> get copyWith =>
-      _$SettingsStateCopyWithImpl<SettingsState>(
-          this as SettingsState, _$identity);
+ bool get progress; int get unreadVoicemailCount;
+/// Create a copy of SettingsState
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$SettingsStateCopyWith<SettingsState> get copyWith => _$SettingsStateCopyWithImpl<SettingsState>(this as SettingsState, _$identity);
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is SettingsState &&
-            (identical(other.progress, progress) ||
-                other.progress == progress) &&
-            (identical(other.unreadVoicemailCount, unreadVoicemailCount) ||
-                other.unreadVoicemailCount == unreadVoicemailCount));
-  }
 
-  @override
-  int get hashCode => Object.hash(runtimeType, progress, unreadVoicemailCount);
 
-  @override
-  String toString() {
-    return 'SettingsState(progress: $progress, unreadVoicemailCount: $unreadVoicemailCount)';
-  }
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is SettingsState&&(identical(other.progress, progress) || other.progress == progress)&&(identical(other.unreadVoicemailCount, unreadVoicemailCount) || other.unreadVoicemailCount == unreadVoicemailCount));
+}
+
+
+@override
+int get hashCode => Object.hash(runtimeType,progress,unreadVoicemailCount);
+
+@override
+String toString() {
+  return 'SettingsState(progress: $progress, unreadVoicemailCount: $unreadVoicemailCount)';
+}
+
+
 }
 
 /// @nodoc
-abstract mixin class $SettingsStateCopyWith<$Res> {
-  factory $SettingsStateCopyWith(
-          SettingsState value, $Res Function(SettingsState) _then) =
-      _$SettingsStateCopyWithImpl;
-  @useResult
-  $Res call({bool progress, int unreadVoicemailCount});
-}
+abstract mixin class $SettingsStateCopyWith<$Res>  {
+  factory $SettingsStateCopyWith(SettingsState value, $Res Function(SettingsState) _then) = _$SettingsStateCopyWithImpl;
+@useResult
+$Res call({
+ bool progress, int unreadVoicemailCount
+});
 
+
+
+
+}
 /// @nodoc
 class _$SettingsStateCopyWithImpl<$Res>
     implements $SettingsStateCopyWith<$Res> {
@@ -62,262 +60,141 @@ class _$SettingsStateCopyWithImpl<$Res>
   final SettingsState _self;
   final $Res Function(SettingsState) _then;
 
-  /// Create a copy of SettingsState
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? progress = null,
-    Object? unreadVoicemailCount = null,
-  }) {
-    return _then(_self.copyWith(
-      progress: null == progress
-          ? _self.progress
-          : progress // ignore: cast_nullable_to_non_nullable
-              as bool,
-      unreadVoicemailCount: null == unreadVoicemailCount
-          ? _self.unreadVoicemailCount
-          : unreadVoicemailCount // ignore: cast_nullable_to_non_nullable
-              as int,
-    ));
-  }
+/// Create a copy of SettingsState
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? progress = null,Object? unreadVoicemailCount = null,}) {
+  return _then(SettingsState(
+progress: null == progress ? _self.progress : progress // ignore: cast_nullable_to_non_nullable
+as bool,unreadVoicemailCount: null == unreadVoicemailCount ? _self.unreadVoicemailCount : unreadVoicemailCount // ignore: cast_nullable_to_non_nullable
+as int,
+  ));
 }
+
+}
+
 
 /// Adds pattern-matching-related methods to [SettingsState].
 extension SettingsStatePatterns on SettingsState {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_SettingsState value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _SettingsState() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>({required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_SettingsState value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _SettingsState():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult map<TResult extends Object?>(){
+final _that = this;
+switch (_that) {
+case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_SettingsState value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _SettingsState() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(){
+final _that = this;
+switch (_that) {
+case _:
+  return null;
 
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(bool progress, int unreadVoicemailCount)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _SettingsState() when $default != null:
-        return $default(_that.progress, _that.unreadVoicemailCount);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>({required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(bool progress, int unreadVoicemailCount) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _SettingsState():
-        return $default(_that.progress, _that.unreadVoicemailCount);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult when<TResult extends Object?>() {final _that = this;
+switch (_that) {
+case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(bool progress, int unreadVoicemailCount)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _SettingsState() when $default != null:
-        return $default(_that.progress, _that.unreadVoicemailCount);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>() {final _that = this;
+switch (_that) {
+case _:
+  return null;
+
+}
 }
 
-/// @nodoc
-
-class _SettingsState extends SettingsState {
-  const _SettingsState({required this.progress, this.unreadVoicemailCount = 0})
-      : super._();
-
-  @override
-  final bool progress;
-  @override
-  @JsonKey()
-  final int unreadVoicemailCount;
-
-  /// Create a copy of SettingsState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  _$SettingsStateCopyWith<_SettingsState> get copyWith =>
-      __$SettingsStateCopyWithImpl<_SettingsState>(this, _$identity);
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _SettingsState &&
-            (identical(other.progress, progress) ||
-                other.progress == progress) &&
-            (identical(other.unreadVoicemailCount, unreadVoicemailCount) ||
-                other.unreadVoicemailCount == unreadVoicemailCount));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, progress, unreadVoicemailCount);
-
-  @override
-  String toString() {
-    return 'SettingsState(progress: $progress, unreadVoicemailCount: $unreadVoicemailCount)';
-  }
-}
-
-/// @nodoc
-abstract mixin class _$SettingsStateCopyWith<$Res>
-    implements $SettingsStateCopyWith<$Res> {
-  factory _$SettingsStateCopyWith(
-          _SettingsState value, $Res Function(_SettingsState) _then) =
-      __$SettingsStateCopyWithImpl;
-  @override
-  @useResult
-  $Res call({bool progress, int unreadVoicemailCount});
-}
-
-/// @nodoc
-class __$SettingsStateCopyWithImpl<$Res>
-    implements _$SettingsStateCopyWith<$Res> {
-  __$SettingsStateCopyWithImpl(this._self, this._then);
-
-  final _SettingsState _self;
-  final $Res Function(_SettingsState) _then;
-
-  /// Create a copy of SettingsState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $Res call({
-    Object? progress = null,
-    Object? unreadVoicemailCount = null,
-  }) {
-    return _then(_SettingsState(
-      progress: null == progress
-          ? _self.progress
-          : progress // ignore: cast_nullable_to_non_nullable
-              as bool,
-      unreadVoicemailCount: null == unreadVoicemailCount
-          ? _self.unreadVoicemailCount
-          : unreadVoicemailCount // ignore: cast_nullable_to_non_nullable
-              as int,
-    ));
-  }
 }
 
 // dart format on
