@@ -1,6 +1,6 @@
 part of 'settings_bloc.dart';
 
-abstract class SettingsEvent {
+sealed class SettingsEvent {
   const SettingsEvent();
 }
 
