@@ -621,7 +621,9 @@ class AppLocalizationsEn extends AppLocalizations {
 
   @override
   String login_CoreVersionUnsupportedExceptionError(
-      String actual, String supportedConstraint) {
+    String actual,
+    String supportedConstraint,
+  ) {
     return 'An incompatible instance version provided, please contact the administrator of your system (actual: $actual, supported: $supportedConstraint)';
   }
 
@@ -809,8 +811,10 @@ class AppLocalizationsEn extends AppLocalizations {
 
   @override
   String
-      main_CompatibilityIssueDialog_contentCoreVersionUnsupportedExceptionError(
-          String actual, String supportedConstraint) {
+  main_CompatibilityIssueDialog_contentCoreVersionUnsupportedExceptionError(
+    String actual,
+    String supportedConstraint,
+  ) {
     return 'Incompatible WebTrit Cloud Backend version, please contact the administrator of your system.\n\nInstance version:\n$actual\n\nSupported version:\n$supportedConstraint';
   }
 
@@ -1157,13 +1161,15 @@ class AppLocalizationsEn extends AppLocalizations {
 
   @override
   String notifications_errorSnackBar_signalingDisconnectWithCodeName(
-      String codeName) {
+    String codeName,
+  ) {
     return 'Disconnected from the core with the code: $codeName';
   }
 
   @override
   String notifications_errorSnackBar_signalingDisconnectWithSystemReason(
-      String reason) {
+    String reason,
+  ) {
     return 'Disconnected from the core due to the following reason: $reason';
   }
 
@@ -1181,7 +1187,8 @@ class AppLocalizationsEn extends AppLocalizations {
 
   @override
   String notifications_errorSnackBar_sipRegistrationFailed_WithSystemReason(
-      String reason) {
+    String reason,
+  ) {
     return 'Registration with the remote VoIP system failed due to the following reason: $reason';
   }
 
@@ -1624,7 +1631,8 @@ class AppLocalizationsEn extends AppLocalizations {
       'Remove static rtpmap lines';
 
   @override
-  String get settings_encoding_Section_extra_sdp_mod_removeStaticRtpmaps_tooltip =>
+  String
+  get settings_encoding_Section_extra_sdp_mod_removeStaticRtpmaps_tooltip =>
       'Remove static RTP map lines for audio codecs (e.g., PCMU, PCMA) from SDP, to reduce SDP size. May help with MTU fragmentation issues on some SIP endpoints.';
 
   @override
