@@ -638,7 +638,9 @@ class AppLocalizationsIt extends AppLocalizations {
 
   @override
   String login_CoreVersionUnsupportedExceptionError(
-      String actual, String supportedConstraint) {
+    String actual,
+    String supportedConstraint,
+  ) {
     return 'È stata fornita una versione di richiesta incompatibile, contattare l\'amministratore del sistema (actual:$actual, supported:$supportedConstraint)';
   }
 
@@ -830,8 +832,10 @@ class AppLocalizationsIt extends AppLocalizations {
 
   @override
   String
-      main_CompatibilityIssueDialog_contentCoreVersionUnsupportedExceptionError(
-          String actual, String supportedConstraint) {
+  main_CompatibilityIssueDialog_contentCoreVersionUnsupportedExceptionError(
+    String actual,
+    String supportedConstraint,
+  ) {
     return 'Versione di WebTrit Cloud Backend incompatibile, si prega di contattare l\'amministratore del sistema.\n\nVersione dell\'istanza:\n$actual\n\nVersione supportata:\n$supportedConstraint\n';
   }
 
@@ -1179,13 +1183,15 @@ class AppLocalizationsIt extends AppLocalizations {
 
   @override
   String notifications_errorSnackBar_signalingDisconnectWithCodeName(
-      String codeName) {
+    String codeName,
+  ) {
     return 'Disconnesso dal core con codice: $codeName';
   }
 
   @override
   String notifications_errorSnackBar_signalingDisconnectWithSystemReason(
-      String reason) {
+    String reason,
+  ) {
     return 'Disconnesso dal nucleo con ragione: $reason';
   }
 
@@ -1203,7 +1209,8 @@ class AppLocalizationsIt extends AppLocalizations {
 
   @override
   String notifications_errorSnackBar_sipRegistrationFailed_WithSystemReason(
-      String reason) {
+    String reason,
+  ) {
     return 'La registrazione con il sistema VoIP remoto è fallita con il motivo: $reason';
   }
 
@@ -1661,7 +1668,8 @@ class AppLocalizationsIt extends AppLocalizations {
       'Rimuovere le linee rtpmap statiche';
 
   @override
-  String get settings_encoding_Section_extra_sdp_mod_removeStaticRtpmaps_tooltip =>
+  String
+  get settings_encoding_Section_extra_sdp_mod_removeStaticRtpmaps_tooltip =>
       'Rimuovere le linee di mappa RTP statiche per i codec audio (ad es. PCMU, PCMA) dall\'SDP, per ridurre la dimensione dell\'SDP. Può aiutare con problemi di frammentazione MTU su alcuni endpoint SIP.';
 
   @override
