@@ -5,17 +5,27 @@ part 'session.freezed.dart';
 part 'session.g.dart';
 
 @freezed
-abstract class Session with _$Session {
-  const Session._();
+@JsonSerializable()
+class Session with _$Session {
+  const Session({
+    this.coreUrl,
+    this.token,
+    this.tenantId = '',
+    this.userId = '',
+  });
 
-  const factory Session({
-    String? coreUrl,
-    String? token,
-    @Default('') String tenantId,
-    @Default('') String userId,
-  }) = _Session;
+  @override
+  final String? coreUrl;
+  @override
+  final String? token;
+  @override
+  final String tenantId;
+  @override
+  final String userId;
 
   factory Session.fromJson(Map<String, dynamic> json) => _$SessionFromJson(json);
+
+  Map<String, dynamic> toJson() => _$SessionToJson(this);
 
   bool get isLoggedIn => coreUrl != null && token != null;
 }
