@@ -14,359 +14,190 @@ T _$identity<T>(T value) => value;
 
 /// @nodoc
 mixin _$Session {
-  String? get coreUrl;
-  String? get token;
-  String get tenantId;
-  String get userId;
 
-  /// Create a copy of Session
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  $SessionCopyWith<Session> get copyWith =>
-      _$SessionCopyWithImpl<Session>(this as Session, _$identity);
+ String? get coreUrl; String? get token; String get tenantId; String get userId;
+/// Create a copy of Session
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$SessionCopyWith<Session> get copyWith => _$SessionCopyWithImpl<Session>(this as Session, _$identity);
 
-  /// Serializes this Session to a JSON map.
-  Map<String, dynamic> toJson();
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is Session &&
-            (identical(other.coreUrl, coreUrl) || other.coreUrl == coreUrl) &&
-            (identical(other.token, token) || other.token == token) &&
-            (identical(other.tenantId, tenantId) ||
-                other.tenantId == tenantId) &&
-            (identical(other.userId, userId) || other.userId == userId));
-  }
 
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  int get hashCode =>
-      Object.hash(runtimeType, coreUrl, token, tenantId, userId);
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is Session&&(identical(other.coreUrl, coreUrl) || other.coreUrl == coreUrl)&&(identical(other.token, token) || other.token == token)&&(identical(other.tenantId, tenantId) || other.tenantId == tenantId)&&(identical(other.userId, userId) || other.userId == userId));
+}
 
-  @override
-  String toString() {
-    return 'Session(coreUrl: $coreUrl, token: $token, tenantId: $tenantId, userId: $userId)';
-  }
+@JsonKey(includeFromJson: false, includeToJson: false)
+@override
+int get hashCode => Object.hash(runtimeType,coreUrl,token,tenantId,userId);
+
+@override
+String toString() {
+  return 'Session(coreUrl: $coreUrl, token: $token, tenantId: $tenantId, userId: $userId)';
+}
+
+
 }
 
 /// @nodoc
-abstract mixin class $SessionCopyWith<$Res> {
-  factory $SessionCopyWith(Session value, $Res Function(Session) _then) =
-      _$SessionCopyWithImpl;
-  @useResult
-  $Res call({String? coreUrl, String? token, String tenantId, String userId});
-}
+abstract mixin class $SessionCopyWith<$Res>  {
+  factory $SessionCopyWith(Session value, $Res Function(Session) _then) = _$SessionCopyWithImpl;
+@useResult
+$Res call({
+ String? coreUrl, String? token, String tenantId, String userId
+});
 
+
+
+
+}
 /// @nodoc
-class _$SessionCopyWithImpl<$Res> implements $SessionCopyWith<$Res> {
+class _$SessionCopyWithImpl<$Res>
+    implements $SessionCopyWith<$Res> {
   _$SessionCopyWithImpl(this._self, this._then);
 
   final Session _self;
   final $Res Function(Session) _then;
 
-  /// Create a copy of Session
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? coreUrl = freezed,
-    Object? token = freezed,
-    Object? tenantId = null,
-    Object? userId = null,
-  }) {
-    return _then(_self.copyWith(
-      coreUrl: freezed == coreUrl
-          ? _self.coreUrl
-          : coreUrl // ignore: cast_nullable_to_non_nullable
-              as String?,
-      token: freezed == token
-          ? _self.token
-          : token // ignore: cast_nullable_to_non_nullable
-              as String?,
-      tenantId: null == tenantId
-          ? _self.tenantId
-          : tenantId // ignore: cast_nullable_to_non_nullable
-              as String,
-      userId: null == userId
-          ? _self.userId
-          : userId // ignore: cast_nullable_to_non_nullable
-              as String,
-    ));
-  }
+/// Create a copy of Session
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? coreUrl = freezed,Object? token = freezed,Object? tenantId = null,Object? userId = null,}) {
+  return _then(Session(
+coreUrl: freezed == coreUrl ? _self.coreUrl : coreUrl // ignore: cast_nullable_to_non_nullable
+as String?,token: freezed == token ? _self.token : token // ignore: cast_nullable_to_non_nullable
+as String?,tenantId: null == tenantId ? _self.tenantId : tenantId // ignore: cast_nullable_to_non_nullable
+as String,userId: null == userId ? _self.userId : userId // ignore: cast_nullable_to_non_nullable
+as String,
+  ));
 }
+
+}
+
 
 /// Adds pattern-matching-related methods to [Session].
 extension SessionPatterns on Session {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_Session value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _Session() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>({required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_Session value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _Session():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult map<TResult extends Object?>(){
+final _that = this;
+switch (_that) {
+case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_Session value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _Session() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(){
+final _that = this;
+switch (_that) {
+case _:
+  return null;
 
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(
-            String? coreUrl, String? token, String tenantId, String userId)?
-        $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _Session() when $default != null:
-        return $default(
-            _that.coreUrl, _that.token, _that.tenantId, _that.userId);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>({required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(
-            String? coreUrl, String? token, String tenantId, String userId)
-        $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _Session():
-        return $default(
-            _that.coreUrl, _that.token, _that.tenantId, _that.userId);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult when<TResult extends Object?>() {final _that = this;
+switch (_that) {
+case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(
-            String? coreUrl, String? token, String tenantId, String userId)?
-        $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _Session() when $default != null:
-        return $default(
-            _that.coreUrl, _that.token, _that.tenantId, _that.userId);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>() {final _that = this;
+switch (_that) {
+case _:
+  return null;
+
+}
 }
 
-/// @nodoc
-@JsonSerializable()
-class _Session extends Session {
-  const _Session(
-      {this.coreUrl, this.token, this.tenantId = '', this.userId = ''})
-      : super._();
-  factory _Session.fromJson(Map<String, dynamic> json) =>
-      _$SessionFromJson(json);
-
-  @override
-  final String? coreUrl;
-  @override
-  final String? token;
-  @override
-  @JsonKey()
-  final String tenantId;
-  @override
-  @JsonKey()
-  final String userId;
-
-  /// Create a copy of Session
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  _$SessionCopyWith<_Session> get copyWith =>
-      __$SessionCopyWithImpl<_Session>(this, _$identity);
-
-  @override
-  Map<String, dynamic> toJson() {
-    return _$SessionToJson(
-      this,
-    );
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _Session &&
-            (identical(other.coreUrl, coreUrl) || other.coreUrl == coreUrl) &&
-            (identical(other.token, token) || other.token == token) &&
-            (identical(other.tenantId, tenantId) ||
-                other.tenantId == tenantId) &&
-            (identical(other.userId, userId) || other.userId == userId));
-  }
-
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  int get hashCode =>
-      Object.hash(runtimeType, coreUrl, token, tenantId, userId);
-
-  @override
-  String toString() {
-    return 'Session(coreUrl: $coreUrl, token: $token, tenantId: $tenantId, userId: $userId)';
-  }
-}
-
-/// @nodoc
-abstract mixin class _$SessionCopyWith<$Res> implements $SessionCopyWith<$Res> {
-  factory _$SessionCopyWith(_Session value, $Res Function(_Session) _then) =
-      __$SessionCopyWithImpl;
-  @override
-  @useResult
-  $Res call({String? coreUrl, String? token, String tenantId, String userId});
-}
-
-/// @nodoc
-class __$SessionCopyWithImpl<$Res> implements _$SessionCopyWith<$Res> {
-  __$SessionCopyWithImpl(this._self, this._then);
-
-  final _Session _self;
-  final $Res Function(_Session) _then;
-
-  /// Create a copy of Session
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $Res call({
-    Object? coreUrl = freezed,
-    Object? token = freezed,
-    Object? tenantId = null,
-    Object? userId = null,
-  }) {
-    return _then(_Session(
-      coreUrl: freezed == coreUrl
-          ? _self.coreUrl
-          : coreUrl // ignore: cast_nullable_to_non_nullable
-              as String?,
-      token: freezed == token
-          ? _self.token
-          : token // ignore: cast_nullable_to_non_nullable
-              as String?,
-      tenantId: null == tenantId
-          ? _self.tenantId
-          : tenantId // ignore: cast_nullable_to_non_nullable
-              as String,
-      userId: null == userId
-          ? _self.userId
-          : userId // ignore: cast_nullable_to_non_nullable
-              as String,
-    ));
-  }
 }
 
 // dart format on
