@@ -10,13 +10,25 @@ part 'raw_http_request.g.dart';
 /// This model is intended for use cases where the request must be defined
 /// independently of internal API clients or repositories.
 @freezed
-abstract class RawHttpRequest with _$RawHttpRequest {
-  const factory RawHttpRequest({
-    required String method,
-    required String url,
-    Map<String, String>? headers,
-    Map<String, dynamic>? data,
-  }) = _RawHttpRequest;
+@JsonSerializable()
+class RawHttpRequest with _$RawHttpRequest {
+  const RawHttpRequest({
+    required this.method,
+    required this.url,
+    this.headers,
+    this.data,
+  });
+
+  @override
+  final String method;
+  @override
+  final String url;
+  @override
+  final Map<String, String>? headers;
+  @override
+  final Map<String, dynamic>? data;
 
   factory RawHttpRequest.fromJson(Map<String, dynamic> json) => _$RawHttpRequestFromJson(json);
+
+  Map<String, dynamic> toJson() => _$RawHttpRequestToJson(this);
 }
