@@ -6,8 +6,8 @@ part of 'raw_http_request.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-_RawHttpRequest _$RawHttpRequestFromJson(Map<String, dynamic> json) =>
-    _RawHttpRequest(
+RawHttpRequest _$RawHttpRequestFromJson(Map<String, dynamic> json) =>
+    RawHttpRequest(
       method: json['method'] as String,
       url: json['url'] as String,
       headers: (json['headers'] as Map<String, dynamic>?)?.map(
@@ -16,7 +16,7 @@ _RawHttpRequest _$RawHttpRequestFromJson(Map<String, dynamic> json) =>
       data: json['data'] as Map<String, dynamic>?,
     );
 
-Map<String, dynamic> _$RawHttpRequestToJson(_RawHttpRequest instance) =>
+Map<String, dynamic> _$RawHttpRequestToJson(RawHttpRequest instance) =>
     <String, dynamic>{
       'method': instance.method,
       'url': instance.url,
